fix:(Shared/pc) 코드리뷰 바탕으로  API호출부분 수정, CSS속성 수정
