+++ src/components/Footer.css
@@ -7,6 +7,9 @@
   margin-top: 40px;
   padding: 32px 104px 62px 104px;
   height: 160px;
+  @media (max-width: 767px) {
+    padding: 32px 32px 64px 32px;
+  }
 }
 
 .footer-container {
@@ -20,9 +23,10 @@
   @media (max-width: 767px) {
     display: grid;
     grid-template-areas:
-      "copyright explain"
-      "sns-logo .";
-    padding: 32px 32px 64px 32px;
+      "explain sns-logo"
+      "copyright .";
+    row-gap: 60px;
+    padding: 0 auto;
   }
 }
