+++ src/components/Nav.js
@@ -1,27 +1,44 @@
 import "./Nav.css";
 import logo from "../assets/logo.svg";
 import { Button } from "../components/Button";
-
-const INIT_USER = {
-  profileImageSource: "",
-  email: "",
-};
-
-const Account = ({ user = INIT_USER }) => {
-  return (
-    <div className="user-account">
-      <img src={user.profileImageSource} alt="프로필 사진" />
-      <div className="user-email">{user.email}</div>
-    </div>
-  );
-};
+import React, { useState, useCallback, useEffect } from "react";
+import { getUserData } from "../services/api";
 
 export default function Nav({ user }) {
+  const [userData, setUserData] = useState([]);
+
+  const INIT_USER = {
+    profileImageSource: "",
+    email: "",
+  };
+
+  const getUserInfo = useCallback(async () => {
+    const userInfo = await getUserData();
+    setUserData(userInfo);
+  });
+
+  useEffect(() => {
+    getUserInfo();
+  }, []); //마운트시점에만 호출
+
+  const Account = ({ user = INIT_USER }) => {
+    return (
+      <div className="user-account">
+        <img src={user.profileImageSource} alt="프로필 사진" />
+        <div className="user-email">{user.email}</div>
+      </div>
+    );
+  };
+
   return (
     <div className="nav-bar">
       <div className="nav-wrapper">
         <img src={logo} alt="로고" className="nav-logo" />
-        {user.email ? <Account user={user} /> : <Button name="로그인" />}
+        {userData.email ? (
+          <Account user={userData} />
+        ) : (
+          <Button name="로그인" />
+        )}
       </div>
     </div>
   );
